Small cosmetic changes to midi-send (deviceId type, remove debug printf)

+++ hardware/tools/mididuino/MidiFirmwareSender.h
@@ -21,7 +21,7 @@ protected:
 
   bool waitingForBootloader;
 
-  int deviceId;
+  uint8_t deviceId;
   int verbose;
   bool statusMessage;
   HexFile *file;
